Remove issued state and caseworker-issue-aos event
