--- src/integrationTest/java/uk/gov/hmcts/divorce/caseworker/CaseworkerIssueAosTest.java
@@ -1,207 +1,0 @@
-package uk.gov.hmcts.divorce.caseworker;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.commons.io.FilenameUtils;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import uk.gov.hmcts.divorce.common.config.WebMvcConfig;
-import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
-import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
-import uk.gov.hmcts.divorce.notification.NotificationService;
-import uk.gov.hmcts.divorce.testutil.DocManagementStoreWireMock;
-import uk.gov.hmcts.divorce.testutil.IdamWireMock;
-import uk.gov.hmcts.divorce.testutil.SendLetterWireMock;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.sendletter.api.LetterStatus;
-import uk.gov.hmcts.reform.sendletter.api.SendLetterResponse;
-
-import java.io.IOException;
-import java.time.Clock;
-import java.time.Instant;
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
-import java.util.List;
-import java.util.UUID;
-
-import static java.util.Arrays.asList;
-import static org.mockito.ArgumentMatchers.anyMap;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoMoreInteractions;
-import static org.mockito.Mockito.when;
-import static org.springframework.http.MediaType.APPLICATION_JSON;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static uk.gov.hmcts.divorce.caseworker.event.CaseworkerIssueAos.CASEWORKER_ISSUE_AOS;
-import static uk.gov.hmcts.divorce.divorcecase.model.LanguagePreference.ENGLISH;
-import static uk.gov.hmcts.divorce.document.model.DocumentType.DIVORCE_APPLICATION;
-import static uk.gov.hmcts.divorce.document.model.DocumentType.DOCUMENT_TYPE_RESPONDENT_INVITATION;
-import static uk.gov.hmcts.divorce.notification.EmailTemplateName.APPLICANT_SOLICITOR_NOTICE_OF_PROCEEDINGS;
-import static uk.gov.hmcts.divorce.notification.EmailTemplateName.RESPONDENT_SOLICITOR_NOTICE_OF_PROCEEDINGS;
-import static uk.gov.hmcts.divorce.testutil.DocManagementStoreWireMock.stubDownloadBinaryFromDocumentManagement;
-import static uk.gov.hmcts.divorce.testutil.IdamWireMock.CASEWORKER_ROLE;
-import static uk.gov.hmcts.divorce.testutil.IdamWireMock.stubForIdamDetails;
-import static uk.gov.hmcts.divorce.testutil.IdamWireMock.stubForIdamToken;
-import static uk.gov.hmcts.divorce.testutil.SendLetterWireMock.stubSendLetters;
-import static uk.gov.hmcts.divorce.testutil.SendLetterWireMock.stubStatusOfSendLetter;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.ABOUT_TO_SUBMIT_URL;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.AUTHORIZATION;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.CASEWORKER_USER_ID;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.SERVICE_AUTHORIZATION;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_AUTHORIZATION_TOKEN;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_SERVICE_AUTH_TOKEN;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_SOLICITOR_EMAIL;
-import static uk.gov.hmcts.divorce.testutil.TestDataHelper.callbackRequest;
-import static uk.gov.hmcts.divorce.testutil.TestDataHelper.documentWithType;
-import static uk.gov.hmcts.divorce.testutil.TestDataHelper.respondentWithDigitalSolicitor;
-import static uk.gov.hmcts.divorce.testutil.TestDataHelper.validCaseDataForIssueApplication;
-import static uk.gov.hmcts.divorce.testutil.TestResourceUtil.expectedResponse;
-import static uk.gov.hmcts.divorce.testutil.TestResourceUtil.resourceAsBytes;
-
-@ExtendWith(SpringExtension.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureMockMvc
-@ContextConfiguration(initializers = {
-    IdamWireMock.PropertiesInitializer.class,
-    DocManagementStoreWireMock.PropertiesInitializer.class,
-    SendLetterWireMock.PropertiesInitializer.class})
-public class CaseworkerIssueAosTest {
-
-    private static final String CASEWORKER_ISSUE_AOS_ABOUT_TO_SUBMIT_REP =
-        "classpath:caseworker-issue-aos-about-to-submit-response.json";
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Autowired
-    private ObjectMapper objectMapper;
-
-    @MockBean
-    private WebMvcConfig webMvcConfig;
-
-    @MockBean
-    private AuthTokenGenerator serviceTokenGenerator;
-
-    @MockBean
-    private NotificationService notificationService;
-
-    @MockBean
-    private Clock clock;
-
-    @BeforeAll
-    static void setUp() {
-        IdamWireMock.start();
-        DocManagementStoreWireMock.start();
-        SendLetterWireMock.start();
-    }
-
-    @AfterAll
-    static void tearDown() {
-        IdamWireMock.stopAndReset();
-        DocManagementStoreWireMock.stopAndReset();
-        SendLetterWireMock.stopAndReset();
-    }
-
-    @Test
-    void shouldSendAosPackAndSetDueDateAndSendNoticeOfProceedingsEmailOnAboutToSubmit() throws Exception {
-
-        final CaseData caseData = validCaseDataForIssueApplication();
-        final Applicant respondent = respondentWithDigitalSolicitor();
-        caseData.setApplicant2(respondent);
-        stubAosPackSendLetter(setupAosPackDocuments(caseData));
-
-        stubForIdamDetails(TEST_AUTHORIZATION_TOKEN, CASEWORKER_USER_ID, CASEWORKER_ROLE);
-        stubForIdamToken(TEST_AUTHORIZATION_TOKEN);
-        setClock();
-
-        mockMvc.perform(MockMvcRequestBuilders.post(ABOUT_TO_SUBMIT_URL)
-            .contentType(APPLICATION_JSON)
-            .header(SERVICE_AUTHORIZATION, TEST_SERVICE_AUTH_TOKEN)
-            .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
-            .content(objectMapper.writeValueAsString(
-                callbackRequest(
-                    caseData,
-                    CASEWORKER_ISSUE_AOS)))
-            .accept(APPLICATION_JSON))
-            .andExpect(
-                status().isOk())
-            .andExpect(
-                content().json(expectedResponse(CASEWORKER_ISSUE_AOS_ABOUT_TO_SUBMIT_REP))
-            );
-
-        verify(notificationService)
-            .sendEmail(
-                eq(TEST_SOLICITOR_EMAIL),
-                eq(APPLICANT_SOLICITOR_NOTICE_OF_PROCEEDINGS),
-                anyMap(),
-                eq(ENGLISH));
-        verify(notificationService)
-            .sendEmail(
-                eq(TEST_SOLICITOR_EMAIL),
-                eq(RESPONDENT_SOLICITOR_NOTICE_OF_PROCEEDINGS),
-                anyMap(),
-                eq(ENGLISH));
-
-        verifyNoMoreInteractions(notificationService);
-    }
-
-    private void stubAosPackSendLetter(List<String> documentIds) throws IOException {
-        final byte[] pdfAsBytes = loadPdfAsBytes();
-
-        for (String documentId : documentIds) {
-            stubDownloadBinaryFromDocumentManagement(documentId, pdfAsBytes);
-        }
-
-        final SendLetterResponse sendLetterResponse = new SendLetterResponse(UUID.randomUUID());
-        final ZonedDateTime now = ZonedDateTime.now();
-        final LetterStatus letterStatus = new LetterStatus(
-            sendLetterResponse.letterId,
-            "OK",
-            "",
-            now,
-            now,
-            now,
-            null,
-            1);
-
-        stubSendLetters(objectMapper.writeValueAsString(sendLetterResponse));
-        stubStatusOfSendLetter(sendLetterResponse.letterId, objectMapper.writeValueAsString(letterStatus));
-
-        when(serviceTokenGenerator.generate()).thenReturn(TEST_SERVICE_AUTH_TOKEN);
-    }
-
-    private List<String> setupAosPackDocuments(final CaseData caseData) {
-        final var documentListValue1 = documentWithType(
-            DIVORCE_APPLICATION,
-            "a2c2c237-866e-44ce-936a-37587c4b4f0c");
-        final var documentListValue2 = documentWithType(
-            DOCUMENT_TYPE_RESPONDENT_INVITATION,
-            "76d34788-4709-4656-ba6a-5bfa1bc17586");
-        final var generatedDocuments = asList(documentListValue1, documentListValue2);
-
-        caseData.setDocumentsGenerated(generatedDocuments);
-
-        return asList(
-            FilenameUtils.getName(documentListValue1.getValue().getDocumentLink().getUrl()),
-            FilenameUtils.getName(documentListValue2.getValue().getDocumentLink().getUrl()));
-    }
-
-    private byte[] loadPdfAsBytes() throws IOException {
-        return resourceAsBytes("classpath:Test.pdf");
-    }
-
-    private void setClock() {
-        when(clock.instant()).thenReturn(Instant.parse("2021-07-06T12:00:00.000Z"));
-        when(clock.getZone()).thenReturn(ZoneId.systemDefault());
-    }
-}
