+++ src/main/java/uk/gov/hmcts/divorce/divorcecase/model/State.java
@@ -29,7 +29,15 @@ public enum State {
         label = "# **${[CASE_REFERENCE]}** ${applicant1LastName} **&** ${applicant2LastName}\n### **${[STATE]}**\n",
         access = {CaseAccessAdministrator.class}
     )
-    AwaitingAos("AwaitingAos"),
+    AwaitingAos("AwaitingAos") {
+        @Override
+        public List<String> validate(CaseData caseData) {
+            List<String> errors = new ArrayList<>();
+            validateBasicCase(caseData, errors);
+            validateCaseFieldsForIssueApplication(caseData.getApplication().getMarriageDetails(), errors);
+            return errors;
+        }
+    },
 
     @CCD(
         name = "AOS drafted",
@@ -54,21 +62,6 @@ public enum State {
         public List<String> validate(CaseData caseData) {
             List<String> errors = new ArrayList<>();
             validateBasicCase(caseData, errors);
-            return errors;
-        }
-    },
-
-    @CCD(
-        name = "Application issued",
-        label = "# **${[CASE_REFERENCE]}** ${applicant1LastName} **&** ${applicant2LastName}\n### **${[STATE]}**\n",
-        access = {CaseAccessAdministrator.class}
-    )
-    Issued("Issued") {
-        @Override
-        public List<String> validate(CaseData caseData) {
-            List<String> errors = new ArrayList<>();
-            validateBasicCase(caseData, errors);
-            validateCaseFieldsForIssueApplication(caseData.getApplication().getMarriageDetails(), errors);
             return errors;
         }
     },
